Onboard workspace demonstrators-social/components
No integration provided, for this reason the build gets broken
