+++ packages/demonstrators/social/components/src/views/ContentView.tsx
